[Fix] 1858 0411 YbH IStatus Playercontroller 등 수정

+++ Assets/YBH/Scripts/IStatus.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 
 public interface IStatus
 {
-    float Health { get; set; }
-    void TakeDamage(float Damage);
+    int Health { get; set; }
+    void TakeDamage(int Damage);
     
     event Action OnDie;
 }
